# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface AuthRegisterRequest {
  /**
     * @minLength 1
     * @maxLength 100
     */
  name: string;
  /** @maxLength 255 */
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface AuthLoginRequest {
  /** @maxLength 255 */
  email: string;
  /** @minLength 1 */
  password: string;
}

export interface AuthTokenClaims {
  id: string;
  email: string;
  iat: number;
  exp: number;
}

export interface AuthUser {
  id: string;
  name: string;
  email: string;
  created_at: string;
}

export interface AuthRegisterResponse {
  success: boolean;
  user: AuthUser;
}

export interface AuthLoginResponse {
  success: boolean;
  token: string;
  user: AuthUser;
}

export interface AuthMeResponse {
  success: boolean;
  user: AuthUser;
}

export interface AuthErrorResponse {
  success: boolean;
  message: string;
}

export interface ProfileCreateRequest {
  education_level?: string;
  institution?: string;
  degree?: string;
  branch?: string;
  graduation_year?: number;
  interests?: string;
  bio?: string;
}

export interface ProfileUpdateRequest {
  education_level?: string;
  institution?: string;
  degree?: string;
  branch?: string;
  graduation_year?: number;
  interests?: string;
  bio?: string;
}

export interface StudentProfile {
  id: string;
  user_id: string;
  /** @nullable */
  education_level: string | null;
  /** @nullable */
  institution: string | null;
  /** @nullable */
  degree: string | null;
  /** @nullable */
  branch: string | null;
  /** @nullable */
  graduation_year: number | null;
  /** @nullable */
  interests: string | null;
  /** @nullable */
  bio: string | null;
  created_at: string;
  updated_at: string;
}

export interface StudentProfileResponse {
  success: boolean;
  profile: StudentProfile;
}

export interface ProfileDeleteResult {
  success: boolean;
  message: string;
}

export interface SupabaseHealthResponse {
  success: boolean;
  message: string;
}

export interface CareerPathHealthResponse {
  success: boolean;
  message: string;
}

export interface HealthStatus {
  status: string;
}