# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: User authentication operations
  - name: profile
    description: Student profile operations
paths:
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a student account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRegisterRequest"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthRegisterResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthErrorResponse"
        "409":
          description: Email is already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthErrorResponse"
        "500":
          description: Registration failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthErrorResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Log in to a student account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthLoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthLoginResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthErrorResponse"
        "401":
          description: Invalid email or password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthErrorResponse"
        "500":
          description: Login failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthErrorResponse"
  /auth/me:
    get:
      operationId: getAuthenticatedUser
      tags: [auth]
      summary: Get the authenticated student
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthMeResponse"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthErrorResponse"
        "500":
          description: Could not load user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthErrorResponse"
  /profile:
    post:
      operationId: createStudentProfile
      tags: [profile]
      summary: Create the authenticated student's profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileCreateRequest"
      responses:
        "201":
          description: Profile created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfileResponse"
        "400":
          description: Invalid profile data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthErrorResponse"
        "409":
          description: A profile already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthErrorResponse"
        "500":
          description: Could not create profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthErrorResponse"
    get:
      operationId: getStudentProfile
      tags: [profile]
      summary: Get the authenticated student's profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Student profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfileResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthErrorResponse"
        "404":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthErrorResponse"
        "500":
          description: Could not load profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthErrorResponse"
    put:
      operationId: updateStudentProfile
      tags: [profile]
      summary: Update the authenticated student's profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdateRequest"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfileResponse"
        "400":
          description: Invalid profile data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthErrorResponse"
        "404":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthErrorResponse"
        "500":
          description: Could not update profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthErrorResponse"
    delete:
      operationId: deleteStudentProfile
      tags: [profile]
      summary: Delete the authenticated student's profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Profile deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileDeleteResult"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthErrorResponse"
        "404":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthErrorResponse"
        "500":
          description: Could not delete profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthErrorResponse"
  /health:
    get:
      operationId: getCareerPathHealth
      tags: [health]
      summary: CareerPath AI API health check
      description: Confirms that the CareerPath AI API is running
      responses:
        "200":
          description: API is running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CareerPathHealthResponse"
  /health/db:
    get:
      operationId: checkSupabaseHealth
      tags: [health]
      summary: Check the Supabase connection
      description: Confirms that the Supabase REST API is reachable using server credentials
      responses:
        "200":
          description: Supabase connection successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupabaseHealthResponse"
        "503":
          description: Supabase connection failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupabaseHealthResponse"
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AuthRegisterRequest:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        email:
          type: string
          format: email
          maxLength: 255
        password:
          type: string
          minLength: 8
      required:
        - name
        - email
        - password
    AuthLoginRequest:
      type: object
      additionalProperties: false
      properties:
        email:
          type: string
          format: email
          maxLength: 255
        password:
          type: string
          minLength: 1
      required:
        - email
        - password
    AuthTokenClaims:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        iat:
          type: integer
        exp:
          type: integer
      required:
        - id
        - email
        - iat
        - exp
    AuthUser:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        created_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - created_at
    AuthRegisterResponse:
      type: object
      properties:
        success:
          type: boolean
        user:
          $ref: "#/components/schemas/AuthUser"
      required:
        - success
        - user
    AuthLoginResponse:
      type: object
      properties:
        success:
          type: boolean
        token:
          type: string
        user:
          $ref: "#/components/schemas/AuthUser"
      required:
        - success
        - token
        - user
    AuthMeResponse:
      type: object
      properties:
        success:
          type: boolean
        user:
          $ref: "#/components/schemas/AuthUser"
      required:
        - success
        - user
    AuthErrorResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
    ProfileCreateRequest:
      type: object
      additionalProperties: false
      properties:
        education_level:
          type: string
        institution:
          type: string
        degree:
          type: string
        branch:
          type: string
        graduation_year:
          type: integer
        interests:
          type: string
        bio:
          type: string
    ProfileUpdateRequest:
      type: object
      additionalProperties: false
      properties:
        education_level:
          type: string
        institution:
          type: string
        degree:
          type: string
        branch:
          type: string
        graduation_year:
          type: integer
        interests:
          type: string
        bio:
          type: string
    StudentProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        education_level:
          type: [string, "null"]
        institution:
          type: [string, "null"]
        degree:
          type: [string, "null"]
        branch:
          type: [string, "null"]
        graduation_year:
          type: [integer, "null"]
        interests:
          type: [string, "null"]
        bio:
          type: [string, "null"]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - user_id
        - education_level
        - institution
        - degree
        - branch
        - graduation_year
        - interests
        - bio
        - created_at
        - updated_at
    StudentProfileResponse:
      type: object
      properties:
        success:
          type: boolean
        profile:
          $ref: "#/components/schemas/StudentProfile"
      required:
        - success
        - profile
    ProfileDeleteResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
    SupabaseHealthResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
    CareerPathHealthResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status