# lib/api-zod/src/generated/types/authLoginRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuthLoginRequest {
  /** @maxLength 255 */
  email: string;
  /** @minLength 1 */
  password: string;
}