# lib/api-zod/src/generated/types/authLoginResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AuthUser } from './authUser';

export interface AuthLoginResponse {
  success: boolean;
  token: string;
  user: AuthUser;
}