# lib/api-zod/src/generated/types/authMeResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AuthUser } from './authUser';

export interface AuthMeResponse {
  success: boolean;
  user: AuthUser;
}