# lib/api-zod/src/generated/types/authRegisterRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuthRegisterRequest {
  /**
     * @minLength 1
     * @maxLength 100
     */
  name: string;
  /** @maxLength 255 */
  email: string;
  /** @minLength 8 */
  password: string;
}