# lib/api-zod/src/generated/types/authRegisterResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AuthUser } from './authUser';

export interface AuthRegisterResponse {
  success: boolean;
  user: AuthUser;
}