# lib/api-zod/src/generated/types/careerPathHealthResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CareerPathHealthResponse {
  success: boolean;
  message: string;
}