# lib/api-zod/src/generated/types/profileCreateRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProfileCreateRequest {
  education_level?: string;
  institution?: string;
  degree?: string;
  branch?: string;
  graduation_year?: number;
  interests?: string;
  bio?: string;
}