# lib/api-zod/src/generated/types/profileDeleteResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProfileDeleteResult {
  success: boolean;
  message: string;
}