# lib/api-zod/src/generated/types/profileUpdateRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProfileUpdateRequest {
  education_level?: string;
  institution?: string;
  degree?: string;
  branch?: string;
  graduation_year?: number;
  interests?: string;
  bio?: string;
}