# lib/api-zod/src/generated/types/studentProfile.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StudentProfile {
  id: string;
  user_id: string;
  /** @nullable */
  education_level: string | null;
  /** @nullable */
  institution: string | null;
  /** @nullable */
  degree: string | null;
  /** @nullable */
  branch: string | null;
  /** @nullable */
  graduation_year: number | null;
  /** @nullable */
  interests: string | null;
  /** @nullable */
  bio: string | null;
  created_at: Date;
  updated_at: Date;
}