# lib/api-zod/src/generated/types/studentProfileResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StudentProfile } from './studentProfile';

export interface StudentProfileResponse {
  success: boolean;
  profile: StudentProfile;
}